Prefetch related revisions for recoverlist_view objects
The base template for recoverlist_view renders fields on the Revision
instance for the Version instances in the queryset. For models with a
large number of deletions, the view becomes too expensive to use.

The typical solution of select_related improves the situation, but in
the case that the number of deleted objects was an issue in the first
place, the number of revisions is also likely large, so the join for
these two tables is expensive (although less expensive than the n+1
queries).

Using prefetch_related results in one more query compare to
select_related, but the aggregate time is much lower when a larger
portion of the objects have been deleted (and similar otherwise).